// galaauction.client/src/types/Item.tsx
export type ItemType = {
    itemId: number;
    itemNumber: number;
    itemName: string;
    winningBidderNumber: number;
    winningBidderName: string;
    winningBidAmount: number;
    isPaid: boolean;
    paymentMethodId: string;
    paymentMethodName: string;
    galaEventId: number;
    categoryId: number;
    categoryName: string;
};